+++ src/Jobs/ImportStages.php
@@ -3,30 +3,30 @@
 namespace Matchish\ScoutElasticSearch\Jobs;
 
 use Illuminate\Support\Collection;
-use Illuminate\Database\Eloquent\Model;
 use Matchish\ScoutElasticSearch\ElasticSearch\Index;
 use Matchish\ScoutElasticSearch\Jobs\Stages\CleanUp;
-use Matchish\ScoutElasticSearch\Jobs\Stages\RefreshIndex;
-use Matchish\ScoutElasticSearch\Jobs\Stages\PullFromSource;
 use Matchish\ScoutElasticSearch\Jobs\Stages\CreateWriteIndex;
+use Matchish\ScoutElasticSearch\Jobs\Stages\PullFromSource;
+use Matchish\ScoutElasticSearch\Jobs\Stages\RefreshIndex;
 use Matchish\ScoutElasticSearch\Jobs\Stages\SwitchToNewAndRemoveOldIndex;
+use Matchish\ScoutElasticSearch\Searchable\ImportSource;
 
 class ImportStages extends Collection
 {
     /**
-     * @param Model $searchable
+     * @param ImportSource $source
      * @return Collection
      */
-    public static function fromSearchable(Model $searchable)
+    public static function fromSource(ImportSource $source)
     {
-        $index = Index::fromSearchable($searchable);
+        $index = Index::fromSource($source);
 
         return (new static([
-            new CleanUp($searchable),
-            new CreateWriteIndex($searchable, $index),
-            PullFromSource::chunked($searchable),
+            new CleanUp($source),
+            new CreateWriteIndex($source, $index),
+            PullFromSource::chunked($source),
             new RefreshIndex($index),
-            new SwitchToNewAndRemoveOldIndex($searchable, $index),
+            new SwitchToNewAndRemoveOldIndex($source, $index),
         ]))->flatten()->filter();
     }
 }