+++ src/Jobs/Stages/CleanUp.php
@@ -3,11 +3,10 @@
 namespace Matchish\ScoutElasticSearch\Jobs\Stages;
 
 use Elasticsearch\Client;
-use Laravel\Scout\Searchable;
-use Illuminate\Database\Eloquent\Model;
 use Elasticsearch\Common\Exceptions\Missing404Exception;
 use Matchish\ScoutElasticSearch\ElasticSearch\Params\Indices\Alias\Get as GetAliasParams;
 use Matchish\ScoutElasticSearch\ElasticSearch\Params\Indices\Delete as DeleteIndexParams;
+use Matchish\ScoutElasticSearch\Searchable\ImportSource;
 
 /**
  * @internal
@@ -15,23 +14,22 @@ use Matchish\ScoutElasticSearch\ElasticSearch\Params\Indices\Delete as DeleteInd
 final class CleanUp
 {
     /**
-     * @var Model
+     * @var ImportSource
      */
-    private $searchable;
+    private $source;
 
     /**
-     * @param Model $searchable
+     * @param ImportSource $source
      */
-    public function __construct(Model $searchable)
+    public function __construct(ImportSource $source)
     {
-        $this->searchable = $searchable;
+        $this->source = $source;
     }
 
     public function handle(Client $elasticsearch): void
     {
-        /** @var Searchable $searchable */
-        $searchable = $this->searchable;
-        $params = GetAliasParams::anyIndex($searchable->searchableAs());
+        $source = $this->source;
+        $params = GetAliasParams::anyIndex($source->searchableAs());
         try {
             /** @var array $response */
             $response = $elasticsearch->indices()->getAlias($params->toArray());