+++ src/Searchable/ImportSource.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Matchish\ScoutElasticSearch\Searchable;
+
+interface ImportSource
+{
+    public function syncWithSearchUsingQueue();
+
+    public function syncWithSearchUsing();
+
+    public function searchableAs();
+
+    public function chunked();
+
+    public function get();
+}