+++ tests/Feature/ImportCommandTest.php
@@ -5,11 +5,12 @@ declare(strict_types=1);
 namespace Tests\Feature;
 
 use App\Book;
-use stdClass;
+use App\BookWithCustomKey;
 use App\Product;
-use Tests\IntegrationTestCase;
 use Illuminate\Support\Facades\Artisan;
+use stdClass;
 use Symfony\Component\Console\Output\BufferedOutput;
+use Tests\IntegrationTestCase;
 
 final class ImportCommandTest extends IntegrationTestCase
 {
@@ -21,6 +22,9 @@ final class ImportCommandTest extends IntegrationTestCase
         $productsAmount = rand(1, 5);
 
         factory(Product::class, $productsAmount)->create();
+
+        $productsUnsearchableAmount = rand(1, 5);
+        factory(Product::class, $productsUnsearchableAmount)->states(['archive'])->create();
 
         Product::setEventDispatcher($dispatcher);
 
@@ -47,7 +51,6 @@ final class ImportCommandTest extends IntegrationTestCase
         Product::unsetEventDispatcher();
 
         $productsAmount = rand(1, 5);
-
         factory(Product::class, $productsAmount)->create();
 
         Product::setEventDispatcher($dispatcher);
@@ -87,6 +90,32 @@ final class ImportCommandTest extends IntegrationTestCase
         ];
         $response = $this->elasticsearch->search($params);
         $this->assertEquals($productsAmount, $response['hits']['total']['value']);
+    }
+
+    public function test_import_with_custom_key_all_pages(): void
+    {
+        $this->app['config']['scout.key'] = 'title';
+
+        $dispatcher = Book::getEventDispatcher();
+        Book::unsetEventDispatcher();
+
+        $booksAmount = 10;
+
+        factory(Book::class, $booksAmount)->create();
+
+        Book::setEventDispatcher($dispatcher);
+
+        Artisan::call('scout:import');
+        $params = [
+            'index' => (new BookWithCustomKey())->searchableAs(),
+            'body' => [
+                'query' => [
+                    'match_all' => new stdClass(),
+                ],
+            ],
+        ];
+        $response = $this->elasticsearch->search($params);
+        $this->assertEquals($booksAmount, $response['hits']['total']);
     }
 
     public function test_remove_old_index_after_switching_to_new(): void