+++ tests/Integration/Jobs/ImportTest.php
@@ -3,11 +3,12 @@
 namespace Tests\Integration\Jobs;
 
 use App\Product;
-use Tests\IntegrationTestCase;
 use Illuminate\Console\OutputStyle;
 use Matchish\ScoutElasticSearch\Jobs\Import;
+use Matchish\ScoutElasticSearch\Searchable\DefaultImportSourceFactory;
 use Symfony\Component\Console\Input\ArrayInput;
 use Symfony\Component\Console\Tests\Fixtures\DummyOutput;
+use Tests\IntegrationTestCase;
 
 class ImportTest extends IntegrationTestCase
 {
@@ -19,7 +20,7 @@ class ImportTest extends IntegrationTestCase
         factory(Product::class, $productsAmount)->create();
         Product::setEventDispatcher($dispatcher);
 
-        $job = new Import(Product::class);
+        $job = new Import(DefaultImportSourceFactory::from(Product::class));
         $output = new DummyOutput();
         $outputStyle = new OutputStyle(new ArrayInput([]), $output);
         $progressBar = $outputStyle->createProgressBar();