+++ tests/Unit/Engines/ElasticSearchEngineTest.php
@@ -3,11 +3,12 @@
 namespace Tests\Unit\Engines;
 
 use App\Product;
-use Mockery as m;
-use Tests\TestCase;
 use Elasticsearch\Client;
 use Laravel\Scout\Builder;
 use Matchish\ScoutElasticSearch\Engines\ElasticSearchEngine;
+use Mockery as m;
+use ONGR\ElasticsearchDSL\Search;
+use Tests\TestCase;
 
 class ElasticSearchEngineTest extends TestCase
 {