+++ CHANGELOG.md
@@ -6,9 +6,23 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 
 ## [Unreleased]
 
-## [2.0.3] - 2019-09-09
+## [3.0.0] - 2019-11-17
 ### Added
+- Elasticsearch 7 support
 - Added interface binding for HitsIteratorAggregate for custom implementation
+
+## [2.1.0] - 2019-11-13
+### Added
+- Import source factory
+- Using global scopes only for import
+
+## [2.0.4] - 2019-11-10
+### Fixed
+- Throw more descriptive exception if there are elasticsearch errors on update
+
+## [2.0.3] - 2019-11-04
+### Fixed
+- Throw exception if there are elasticsearch errors on update
 
 ## [2.0.2] - 2019-05-10
 ### Added